Refactor Page For Better Organization
