+++ packages/bodiless-page/src/Forms/index.ts
@@ -1,13 +1,6 @@
-import withClonePageButton from './withClonePageButton';
-import withDeletePageButton from './withDeletePageButton';
-import withMovePageButton from './withMovePageButton';
-import withNewPageButton from './withNewPageButton';
-
-export * from './PageForm';
-export * from './PageURLField';
-export {
-  withClonePageButton,
-  withDeletePageButton,
-  withMovePageButton,
-  withNewPageButton,
-};
+export * from './MenuFormFields';
+export * from './MenuFormPage';
+export * from './MenuFormPageClone';
+export * from './MenuFormPageDelete';
+export * from './MenuFormPageMove';
+export * from './MenuFormPageNew';
