+++ packages/bodiless-page/src/MenuOptions/index.ts
@@ -1,0 +1,4 @@
+export * from './withClonePageButton';
+export * from './withDeletePageButton';
+export * from './withMovePageButton';
+export * from './withNewPageButton';
